히스토리 post, get api 수정( 유저 아이디 잘못됨 )

+++ src/main/java/com/kakaobank/booksearch/service/HistoryService.java
@@ -21,7 +21,7 @@ public class HistoryService {
     }
 
     @Transactional
-    public void postHistory(int userId, String query) {
+    public void postHistory(long userId, String query) {
         History history = new History();
         history.setUserId(userId);
         history.setQuery(query);
